+++ pkg/label/label.go
@@ -3,4 +3,7 @@ package label
 const (
 	Cluster      = "giantswarm.io/cluster"
 	Organization = "giantswarm.io/organization"
+	// Labels, used in legacy cluster namespaces
+	LegacyCluster  = "cluster"
+	LegacyCustomer = "customer"
 )